+++ packages/app-webdir-ui/src/ProfileCard/index.js
@@ -98,7 +98,9 @@ const ProfileCard = ({ ...props }) => {
         )}
         {!props.profileURL && <h3 className="person-name">{props.name}</h3>}
         <div className="person-profession">
-          <h4>{title}</h4>
+          <p className="person-profession-title">
+            <strong>{title}</strong>
+          </p>
           {department && <span className="dept">{department}</span>}
         </div>
         {props.size !== "micro" && (