+++ packages/component-header-footer/src/header/index.stories.js
@@ -121,7 +121,7 @@ Default.args = {
   loggedIn: false,
   userName: "",
   navTree: basicNavTree,
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   breakpoint: "Lg",
   searchUrl: "https://search.asu.edu/search",
   site: "subdomain",
@@ -133,7 +133,7 @@ Empty.args = {};
 
 export const NoNavigation = Template.bind({});
 NoNavigation.args = {
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   loggedIn: true,
   userName: "Sparky",
   logoutLink: "/caslogout",
@@ -144,7 +144,7 @@ NoNavigation.args = {
 
 export const NoNavigationWithButtons = Template.bind({});
 NoNavigationWithButtons.args = {
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   buttons: [
     {
       href: "/",
@@ -172,7 +172,7 @@ BreakpointXL.args = {
   logoutLink: "/caslogout",
   loginLink: "/cas",
   navTree: basicNavTree,
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   parentOrg: "Parent unit name",
   parentOrgUrl: "https://engineering.asu.edu",
   breakpoint: "Xl",
@@ -188,7 +188,7 @@ WithMobileNavTree.args = {
   loginLink: "/cas",
   navTree: basicNavTree,
   mobileNavTree,
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   parentOrg: "Parent unit name",
   parentOrgUrl: "https://engineering.asu.edu",
   breakpoint: "Xl",
@@ -199,7 +199,7 @@ WithMobileNavTree.args = {
 export const WithButtons = Template.bind({});
 WithButtons.args = {
   navTree: navTreeWithButtons,
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   buttons: [
     {
       href: "/",
@@ -224,7 +224,7 @@ WithButtons.args = {
 export const WithMenuColumns = Template.bind({});
 WithMenuColumns.args = {
   navTree: navTreeMega,
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   parentOrg: "Parent unit name",
   parentOrgUrl: "https://engineering.asu.edu",
   loggedIn: true,
@@ -238,7 +238,7 @@ WithMenuColumns.args = {
 export const ExpandOnHover = Template.bind({});
 ExpandOnHover.args = {
   navTree: navTreeMega,
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   parentOrg: "Parent unit name",
   parentOrgUrl: "https://engineering.asu.edu",
   loggedIn: true,
@@ -276,7 +276,7 @@ Unbranded.args = {
 
 export const AnimatedTitle = AnimatedTitleTemplate.bind({});
 AnimatedTitle.args = {
-  title: "Subdomain name",
+  title: "Subdomain name Lorem ipsum dolor sit amet, consectetur adipiscing elit",
   navTree: basicNavTree,
   loggedIn: false,
   logoutLink: "/caslogout",